Move player feature into route group
