+++ app/(features)/player/components/Timeline.tsx
