+++ app/(features)/player/components/TrackInfo.tsx
