+++ app/features/juz/__tests__/JuzPage.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
 import JuzPage from '@/app/features/juz/[juzId]/page';
