+++ app/features/page/__tests__/PagePage.test.tsx
@@ -1,6 +1,6 @@
 import { render, act } from '@testing-library/react';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
 import QuranPage from '@/app/features/page/[pageId]/page';
