+++ app/features/page/layout.tsx
@@ -1,6 +1,6 @@
 // app/features/page/layout.tsx
 'use client';
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 
 export default function PageLayout({ children }: { children: React.ReactNode }) {
   return <AudioProvider>{children}</AudioProvider>;
