+++ app/features/surah/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 
 export default function SurahLayout({ children }: { children: React.ReactNode }) {
   return <AudioProvider>{children}</AudioProvider>;
