+++ app/features/tafsir/[surahId]/[ayahId]/components/VerseCard.tsx
@@ -2,7 +2,7 @@
 import { FaPlay, FaPause, FaBookmark, FaRegBookmark, FaShare } from '@/app/shared/SvgIcons';
 import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/text/languageCodes';
-import { useAudio } from '@/app/features/player/context/AudioContext';
+import { useAudio } from '@/app/(features)/player/context/AudioContext';
 import { useSettings } from '@/app/providers/SettingsContext';
 import Spinner from '@/app/shared/Spinner';
 import { applyTajweed } from '@/lib/text/tajweed';
