+++ app/features/tafsir/__tests__/TafsirVersePage.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import TafsirVersePage from '@/app/features/tafsir/[surahId]/[ayahId]/page';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
 import { Verse } from '@/types';
