+++ app/features/tafsir/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
 // app/features/tafsir/layout.tsx
-import { AudioProvider } from '@/app/features/player/context/AudioContext';
+import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
 
 export default function TafsirLayout({ children }: { children: React.ReactNode }) {
   return <AudioProvider>{children}</AudioProvider>;
